requirements_final.txt
absl-py==1.2.0
alembic==1.10.3
asttokens==2.0.8
attrs==23.1.0
autoimpute==0.11.7
backcall==0.2.0
backports.functools-lru-cache==1.6.4
blackjax==0.9.6
bokeh==2.4.3
brotlipy==0.7.0
certifi==2023.11.17
cffi==1.14.6
chex==0.1.5
click==8.1.3
cloudpickle==2.2.0
clustree==0.2.1
cmaes==0.9.1
cmcrameri==1.4
colorama==0.4.6
colorlog==6.7.0
comm==0.1.4
commonmark==0.9.1
contourpy==1.0.6
cryptography==38.0.3
cycler==0.11.0
cytoolz==0.12.0
dask==2022.5.2
dask-jobqueue==0.7.4
debugpy==1.6.3
decorator==5.1.1
distributed==2022.5.2
dm-tree==0.1.7
dynamax==0.1.0+197.g0cde118
entrypoints==0.4
esi-acme==2022.8
etils==0.9.0
executing==1.1.0
farrow-and-ball==0.0.3
fastjsonschema==2.17.1
fastprogress==1.0.3
flatbuffers==2.0
flax==0.6.1
fonttools==4.38.0
fsspec==2022.11.0
gast==0.5.3
greenlet==2.0.2
h5py==3.7.0
HeapDict==1.0.1
hmmlearn==0.2.8
idna==3.4
igraph==0.10.4
imbalanced-learn==0.9.1
importlib-metadata==6.5.0
importlib-resources==5.10.0
ipykernel==6.20.2
ipython==8.5.0
jax==0.3.23
jaxlib==0.3.22
jaxopt==0.5.5
jaxtyping==0.2.7
jedi==0.18.1
Jinja2==3.1.2
joblib==1.2.0
joypy==0.2.6
jsonschema==4.17.3
jupyter-client==7.3.4
jupyter_core==4.11.1
kiwisolver==1.4.4
kneed==0.8.1
llvmlite==0.40.0
locket==1.0.0
lz4==4.0.2
Mako==1.2.4
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
matplotx==0.3.10
missingno==0.5.1
msgpack==1.0.4
munkres==1.1.4
natsort==8.2.0
nbformat==5.8.0
nest-asyncio==1.5.6
networkx==3.1
numba==0.57.0
numpy==1.23.4
opencv-python==4.7.0.72
opt-einsum==3.3.0
optax==0.1.3
optuna==3.1.1
packaging==21.3
pairing-functions==0.2.1
pandas==1.5.1
parso==0.8.3
partd==1.3.0
patsy==0.5.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
plotly==5.14.1
prompt-toolkit==3.0.31
psutil==5.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.13.0
pymc3==3.11.5
pynndescent==0.5.10
pyOpenSSL==22.1.0
pyparsing==3.0.9
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.5
PyYAML==6.0
pyzmq==24.0.1
rich==12.6.0
scikit-learn==1.1.3
scipy==1.9.3
seaborn==0.11.2
semver==3.0.0.dev4
six==1.16.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.9
stack-data==0.5.1
statsmodels==0.13.2
style==1.1.0
tblib==1.7.0
tenacity==8.2.2
tensorflow-probability==0.18.0
texttable==1.6.7
Theano-PyMC==1.1.2
threadpoolctl==3.1.0
toolz==0.12.0
tornado==6.1
tqdm==4.65.0
traitlets==5.4.0
tslearn==0.6.2
typeguard==2.13.3
typing_extensions==4.4.0
umap==0.1.1
umap-learn==0.5.3
unicodedata2==15.0.0
update==0.0.1
urllib3==1.26.12
wcwidth==0.2.5
xgboost==1.7.3
xyzservices==2022.9.0
zict==2.2.0
zipp==3.10.0